requirements.txt
Django
django-admin-sortable
django-markdownfield
django-phonenumber-field
djangorestframework
django-crispy-forms
django-localflavor
mysqlclient
phonenumbers
gunicorn
